requirements.txt
mysql-connector-python
flask-sqlalchemy
sqlalchemy
pymysql
PyJWT
google-cloud-storage
numpy==1.23.4
tensorflow==2.12.0
opencv-python==4.5.3.56
python-dotenv
pillow